Move `cf` logic into `cf-to-llvm` pass
