+++ xrcf/src/convert/mod.rs
@@ -12,11 +12,13 @@ use std::sync::Arc;
 use std::sync::RwLock;
 use tracing::debug;
 
+mod cf_to_llvm;
 mod experimental_to_mlir;
 mod func_to_llvm;
 mod mlir_to_llvmir;
 mod scf_to_cf;
 
+pub use cf_to_llvm::ConvertCFToLLVM;
 pub use experimental_to_mlir::ConvertExperimentalToMLIR;
 pub use func_to_llvm::ConvertFuncToLLVM;
 pub use mlir_to_llvmir::ConvertMLIRToLLVMIR;
