+++ xrcf/tests/convert/mod.rs
@@ -1,3 +1,4 @@
+mod cf_to_llvm;
 mod experimental_to_mlir;
 mod func_to_llvm;
 mod mlir_to_llvmir;
